adjoint/examples/embeding_inversion/requirements.txt
diffusers>==0.5.0
accelerate
torchvision
transformers>=4.21.0
ftfy
tensorboard
modelcards
torchdiffeq